requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
python-dotenv==1.0.0
cloudinary==1.36.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
gunicorn
flask-session==0.5.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3  # Better parser for BeautifulSoup
Pillow==10.0.0
google-generativeai==0.3.2 
linkedin-api==2.0.0a4
oauthlib==3.2.2
requests-oauthlib==1.3.1
tweepy==4.14.0